Move document constants into separate file
