+++ src/main/steps/application-withdrawn/get.test.ts
@@ -24,7 +24,7 @@ describe('WithdrawApplicationController', () => {
           userCase: req.session.userCase,
         }),
         ...defaultViewArgs,
-        isAmendableStates: false,
+        isAmendableStates: undefined,
       })
     );
   });
