+++ src/main/steps/contact-us/get.test.ts
@@ -24,7 +24,7 @@ describe('ContactUsGetController', () => {
           userCase: req.session.userCase,
         }),
         ...defaultViewArgs,
-        isAmendableStates: false,
+        isAmendableStates: undefined,
       })
     );
   });
