+++ src/main/steps/terms-and-conditions/get.test.ts
@@ -24,7 +24,7 @@ describe('TermsAndConditionsGetController', () => {
           userCase: req.session.userCase,
         }),
         ...defaultViewArgs,
-        isAmendableStates: false,
+        isAmendableStates: undefined,
       })
     );
   });
