+++ src/main/steps/timed-out/get.test.ts
@@ -18,7 +18,7 @@ describe('TimedOutGetController', () => {
     expect(res.render).toHaveBeenCalledWith(expect.anything(), {
       ...controller.getPageContent(req, res, language),
       ...defaultViewArgs,
-      isAmendableStates: false,
+      isAmendableStates: undefined,
     });
   });
 });
